Bugfixing & Adding more functions to controllers
