+++ src/main/java/com/spring/demo/controllers/ProductController.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequestMapping(path = "/products")
 public class ProductController {
@@ -30,9 +32,15 @@ public class ProductController {
     }
 
     @GetMapping("/searchByName")
-    public Product searchProduct(@RequestParam String name){
-        return repo.getProductByName(name).orElseThrow( () -> new ProductNotFoundException("name" + name));
+    public List <Product> searchProduct(@RequestParam String name) {
+        try {
+            return repo.getProductByName(name);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new ProductNotFoundException("name" + name);
+        }
     }
+
 
     @PostMapping("/create")
     public Product createProduct(@RequestBody Product prod){
