+++ src/main/java/com/spring/demo/models/Employee.java
@@ -5,18 +5,25 @@ import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
+import java.beans.Transient;
 import java.util.List;
 
 @Entity
 @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
 @Table(name = "employee")
-public class Employee extends Human {
+public class Employee {
 
     @Id
     @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
     @GeneratedValue(generator = "uuid2")
     @Column(length = 64)
     private String id;
+
+    private String firstName;
+    private String lastName;
+    @Column(unique = true)
+    private String email;
+    private String phoneNumber;
 
     private String field;
 
@@ -45,5 +52,49 @@ public class Employee extends Human {
 
     public void setShipment(List<Shipment> shipment) {
         this.shipment = shipment;
+    }
+
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    @Transient
+    public String getName(){
+        StringBuilder name = new StringBuilder();
+
+        name.append(firstName);
+        name.append(" ");
+        name.append(lastName);
+
+        return name.toString();
     }
 }
