--- src/main/java/com/spring/demo/models/Human.java
@@ -1,57 +1,0 @@
-package com.spring.demo.models;
-
-import javax.persistence.Column;
-import java.beans.Transient;
-
-public class Human {
-
-    private String firstName;
-    private String lastName;
-    @Column(unique = true)
-    private String email;
-    private String phoneNumber;
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-    @Transient
-    public String getName(){
-        StringBuilder name = new StringBuilder();
-
-        name.append(firstName);
-        name.append(" ");
-        name.append(lastName);
-
-        return name.toString();
-    }
-
-}
