+++ src/main/java/com/spring/demo/models/Product.java
@@ -1,6 +1,7 @@
 package com.spring.demo.models;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import org.hibernate.annotations.GenericGenerator;
 
@@ -24,6 +25,7 @@ public class Product {
     @ManyToOne
     private Manufacturer manufacturer;
     @ManyToMany()
+    @JsonIgnore
     private List<Tag> tags;
 
     public String getId(){
