+++ src/main/java/com/spring/demo/models/User.java
@@ -4,16 +4,23 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
+import java.beans.Transient;
 
 @Entity
 @Table(name = "users")
-public class User extends Human {
+public class User {
 
     @Id
     @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
     @GeneratedValue(generator = "uuid2")
     @Column(length = 64)
     private String id;
+
+    private String firstName;
+    private String lastName;
+    @Column(unique = true)
+    private String email;
+    private String phoneNumber;
 
     @OneToOne(mappedBy = "user", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JsonIgnore
@@ -32,5 +39,49 @@ public class User extends Human {
 
     public void setCart(ShoppingCart cart) {
         this.cart = cart;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+
+    @Transient
+    public String getName(){
+        StringBuilder name = new StringBuilder();
+
+        name.append(firstName);
+        name.append(" ");
+        name.append(lastName);
+
+        return name.toString();
     }
 }
