+++ src/main/java/com/spring/demo/repositories/EmployeeRepository.java
@@ -1,7 +1,14 @@
 package com.spring.demo.repositories;
 
 import com.spring.demo.models.Employee;
+import com.spring.demo.models.User;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.Optional;
+
 public interface EmployeeRepository extends CrudRepository<Employee, String> {
+
+    @Query(value = "select employee from Employee employee where employee.firstName like %?1")
+    Optional<Employee> getEmployeeByName(String firstName);
 }
